Move docker progress reader
